# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AyuTrace Backend API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing ayurvedic herbs supply chain traceability system",
    "contact": {
      "name": "AyuTrace Team",
      "email": "support@ayutrace.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.ayutrace.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "SupplyChainEventType": {
        "type": "string",
        "enum": [
          "PROCESSING",
          "TESTING",
          "TRANSFER",
          "STORAGE",
          "PACKAGING"
        ],
        "description": "Type of supply chain event",
        "example": "PROCESSING"
      },
...
    }
  },
  "paths": {
...
    "/api/supply-chain": {
      "post": {
        "summary": "Create a new supply chain event",
        "description": "Records a new event in the supply chain tracking system",
        "tags": [
          "Supply Chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "eventType",
                  "handlerId",
                  "fromLocationId",
                  "toLocationId"
                ],
                "properties": {
                  "eventType": {
                    "$ref": "#/components/schemas/SupplyChainEventType"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Event timestamp (defaults to current time)",
                    "example": "2025-09-05T10:30:00Z"
                  },
                  "rawMaterialBatchId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Associated raw material batch ID (optional)",
                    "example": "c1d2e3f4-g5h6-7890-1234-567890abcdef"
                  },
                  "finishedGoodId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Associated finished good ID (optional)",
                    "example": "e1f2g3h4-i5j6-7890-1234-567890abcdef"
                  },
                  "handlerId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the user handling the event",
                    "example": "b1c2d3e4-f5g6-7890-1234-567890abcdef"
                  },
                  "fromLocationId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the source organization/location",
                    "example": "f1g2h3i4-j5k6-7890-1234-567890abcdef"
                  },
                  "toLocationId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the destination organization/location",
                    "example": "g1h2i3j4-k5l6-7890-1234-567890abcdef"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional notes about the event",
                    "example": "Quality check passed, transferred to processing facility"
                  },
                  "custody": {
                    "type": "object",
                    "description": "Custody information as JSON",
                    "example": {
                      "temperature": "20°C",
                      "humidity": "45%",
                      "transportMethod": "refrigerated truck"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Supply chain event created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all supply chain events",
        "description": "Retrieve supply chain events with pagination and filtering",
        "tags": [
          "Supply Chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "eventType",
            "schema": {
              "$ref": "#/components/schemas/SupplyChainEventType"
            },
            "description": "Filter by event type"
          },
          {
            "in": "query",
            "name": "batchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by batch ID"
          },
          {
            "in": "query",
            "name": "handlerId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by handler ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of supply chain events"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}